Add initial implementation of ID type
This was largely pulled from a different project into this new library
so it can be reused.

The goal here is to provide a bridge between `github.com/google/uuid`
and `github.com/lithammer/shortuuid` while also supporting database
interfaces so they can continue to be used as column types in things
like `gorm.io/gorm`.